requirements.txt
pandas
python-dotenv
llama-index
openpyxl
chromadb
openai
streamlit 